Updated
Updated the navbar to hide when in mobile view, updated the buttons to be completely invisible (will have to show a selected state but we can worry about that later). Abstracted the header into a separate component so we can reuse it across dashabord and leaderboard. The changes were reviewed with Alisha already.