Refactored the backend, still some things to clean up
Reorganized the backend as if it was a 'microservices' architecture. All the website code is now in the website/ folder, and inside of that folder is separated by separate topics/groups of functions. For example, leetcode/ now contains all the logic for itself to function, such as repo/, model/, client/. client/ specifically can stand on it's own and be reused across these group boundaries. There is still some work to do, but it is a good point to leave it off.