Update Neovim config & make Checkstyle and Java Formatter have consistent rules
The biggest problem with the old .nvim.lua file is that it does not work for lazy.nvim plugins whatsoever. It was essentially a race condition hack that would fail half of the time. However, this should no longer be an issue as the .lazy.lua file will override any plugins that have been set inside of the base neovim config. This will now properly apply the proper plugin overrides required to use Checkstyle and Java formatter within Neovim. .nvim.lua still enforces the tab size to be 4 spaces. Improved Checkstyle and Java Formatter config files to be more consistent.