Abstract submissions per user into separate routes (DESC)
Querying submissions and attaching an user object from within that SQL query has been very brittle and led to very weird SQL queries (especially when trying to implement filters on the question). As a result, this is now separated into two endpoints, fetching a user profile, and fetching user submissions. These queries are also moved out of the SubmissionController as they tie more to finding data on any user, compared to the actual heavier logic (that is more tied to Leetcode.com) in SubmissionController.