Implement user tags (DESC)
Implemented User Tags which is a table that is derived from the User table. All User Tags are automatically jointed to the User models from the SQL repositories, but in order to update User Tags, you have to use the UserTagRepository. These tags will allow us to mark/filter for Patina members, but we can also trivially add support for universities with college emails. This hasn't actually been implemented into the core logic yet, just the implementation.