Implemented new leaderboard logic/route to frontend (DESC)
Implemented the slightly different leaderboard routes, as well as remove authentication for all leaderboard and submission routes on the backend and frontend. Some types on the frontend are not updated to better match the DTO. TODO: Abstract some of these hooks and delete some of the hooks that are not used.