Enabled asset caching on the server (DESC)
This will stop the bug where the logo image is changed every single time on the frontend. The TTL is set relatively low at 30m so that if we decide to change the image or logo, it will expire relatively soon. This won't affect .js,.css,.html or any other files from Vite because those files have hashes and change everytime we re-deploy.