Fix useUrlState bug with page in leaderboard query (DESC)
The problem was that the useEffect to reset the page to page 1 when patina or query changes would cause the page state to break because it would be toggled back to page 1 before the state would finish running. The better solution would be to reset the page whenever the actual functions to update the patina filter or updagte query actually runs, which will only run when invoked, not on render.