Added logic to deactivate current leaderboard in the leaderboard create process (DESC)
This is a very powerful route, so we should have multiple levels of confirmation on the frontend (similar to how GitHub makes you write out the full name of the repository) to ensure that this operation is not accidental. The reason for this is to avoid race conditions from separating the routes. Even if for a couple seconds, users can navigate to our site and error due to no active leaderboard.