Implemented Discord message on leaderboard completion (DESC)
Codebloom now sends a Discord message when an admin creates a brand new leaderboard. This can be a little slow, so there is now a loader on the new leaderboard form as well as a notifcation that says please wait, this may take a while. Furthermore, abstracted the JDA Client so that the rich embed message sending can be reused for both the weekly scheduled message as well as within the controller when the leaderboard is completed