Add openapi-typescript and setup files + generation command (DESC)
Setup a generate script as well as a generate:dev script that will reload all types if the backend files change in development mode. This new script runs in make dev via concurrently, so it should be a drop-in replacement. The new lib/api/types/autogen/schema.d.ts file is where these automatically generated types will live, but we will still manually dedupe the types and redefine them manually for use in our codebase.