Add a hack to fix modal not updating on success (DESC)
Even though invalidateQueries is an async function, it is blocking the execution of the mutate() hook level onSuccess, which is blocking the notification message update to show the response from the server. An acceptable hack is to wrap invalidateQueries in a setTimeout, and defer it's execution via a callback 100ms from the register point.