Update the schema generator file to use .ts instead of .d.ts (DESC)
There was a bug that was happening when attempting to export an imported enum from a .d.ts file, which would cause vite's import resolver to crash consistently. The issue might be that .d.ts files should be ambient types, but enums are not erasable types and so Vite's bundler doesn't know exactly how to handle it. The easiest fix was to convert the schema file to a regular .ts file, which seems to make Vite happy now.