Fixed a bug where  was not correctly stopping threads from doing non-atomic/safe things. (DESC)
The bug was that threads in different pools would not respect the  keyword and would try to read the cookie when it's not supposed to. For example, thread 1 from pool 1 would see that cookie is not available in db (and redis if ci) and enter the long process of fetching. Previously, when pool 2 thread 1 would not wait and try to read the cookie before the previous pool's thread is done, and so then this thread would enter the long process of fetching as well. This is obviously an annoying issue (controllers use 1 pool from what I recall, but it does not share that pool with Async or Scheduled methods which have their own pool) and so the fix would be to use a ReentrantLock (specifically a rw lock since we do infrequent writes and many reads at once) and make that lock static so it is shared ACROSS thread pools (it will only be initialized once when the class is loaded into the JVM). This now fixes the issue, but as a result, the syntax required to use a lock is a bit \'uglier\'. It also required there to be better boundaries between when the actual cookie is being fetched from leetcode vs the async method on start that checks the database for tokens.