Fix bug where leaderboard was resetting to 1 on page load (DESC)
The issue was that on initial render, a useEffect would run that would force the page back to 1. The behavior was unintended because that effect was only supposed to run any time a filter would change (to prevent page out-of-bounds). A possible solution was to use a ref, but it's somewhat clunky. The better solution was to implement a function that can be passed into useFilter that can run as a side-effect whenever a filter would be toggled.

merge conflicts

undo unnecessary code changes