493: Add shared logic in DuelManager.endDuel, use in manual controller method & in auto cleanup duel service (DESC)
Add shared method DuelController.endDuel which can possibly throw a DuelException. This DuelException class can contain an HttpStatus, which is useful for any HTTP-facing logic (like controllers) but can be ignored by background services. Added new methods to LobbyRepository in order to query the necessary lobbies required for this feature.